requirements.txt
aiohttp==3.10.10
aiosignal==1.3.1
APScheduler==3.10.4
attrs==24.2.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.0
six==1.16.0
tzlocal==5.2
yarl==1.17.1
